pyproject.toml
[project]
name = "papers-digest-ai"
version = "0.1.0"
description = "Daily AI agent that finds, ranks, and posts today's papers."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "python-dateutil>=2.9.0",
  "feedparser>=6.0.11",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
papers-digest = "papers_digest.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

